Rearrange test scripts
Tests work for non-MPI builds.

Increased the default per-test timeout because it was too small for the long tests.

Use a more compact description of the scalar checks.
Add logic so that if a QMC run is not added, subsequent scalar checks are also
not added.

Added an output variable to RUN_QMC_APP so calling code can determine
if the test was added or not (due to to number of MPI processes)


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6757 e5b18d87-469d-4833-9cc0-8cdfa06e9491