Fix use of TEST_ADDED argument to RUN_QMC_APP.
Output arguments to functions need to be dereferenced before setting them.
The previous code worked because the caller used the same name (TEST_ADDED)
as in the function code.  The SET command was directly setting the caller's
variable (eg, the TEST_ADDED parameter could have been removed from the RUN_QMC_APP
definition and it would still work, which is not the behavior we want.)

With the change, the function is actually setting the variable that is
given as a parameter in the caller.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6758 e5b18d87-469d-4833-9cc0-8cdfa06e9491