Add pass/fail check for basic wavefunction test.
The basic wavefunction test (runBasicTest) has these checks:
- finite different gradient and Laplacian
    - single delta value against a single tolerance
    - ensure error decreases with a sequence of decreasing delta values
- ratio for trial move

The initial tolerances may be appropriate for double precision, but do fail for
single precision wavefunctions.  Should get the target precision
of the wavefunction and adjust these values appropriately.

Plotting the error vs. delta is helpful in understanding whether deviations are
a finite difference/precision issue or an actual code issue.  There is code to
output this data, but it is off by default.  It might be useful to add an
input file option to turn this on.

Other TODO items:
 - test wavefunction components separately
 - add check for complex wavefunction ratio


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6765 e5b18d87-469d-4833-9cc0-8cdfa06e9491