Use Richardson extrapolation to compute finite difference derivatives.
This lessens the dependence on the choice of delta, and should make it
more reliable to compare with the analytic derivative.

Remove the test using a sequence of delta values, as it is not needed.

Move a number of the routines related to finite differences to a separate class.

What did not seem to work:
 - Using a higher order formula for the derivatives.  It had lower error for large
   delta, but ran into numerical problems sooner.  The problem of choosing the
   correct value for delta remained.
 - Computing higher order derivatives to estimate the rough size of the error term
   in the finite difference formula and to provide a guide for the right tolerance
   for comparison.  Unfortunately, these are even more sensitive to the right
   choice of delta.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6795 e5b18d87-469d-4833-9cc0-8cdfa06e9491