Fix hang in converting charged systems.
The set of lone charges (loneQ) is collected atom-by-atom.
In a positively charged system, not all of them will actually be present.
Change the while loop to terminate if all the electrons have been assigned.

Add a test case and a 30 second timeout to the converter tests.

(Failing example came from Paul Yang)


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6810 e5b18d87-469d-4833-9cc0-8cdfa06e9491