Add input parameters for finite difference wavefunction checks.
The finite difference step size (delta) can be adjusted,
as can the relative tolerance used for the pass/fail decision.
Example:
   <qmc method="wftest">
     <delta>0.01</delta>
     <tolerance>0.1</tolerance>
   </qmc>

Especially for single precision splines, the delta may need to be lowered
from the default value (0.0001).

The relative error as a function of the step size is not output by default,
but can be enabled using the "delta_output" element.

Example:
   <qmc method="wftest">
        <delta_output>
            <file>delta2.dat</file>
            <particle_index>1</particle_index>
            <gradient_index>0</gradient_index>
        </delta_output>
    </qmc>

The "file", "particle_index", and "gradient_index" parameters are all optional.
The default file name is "delta.dat", the default particle index is 0, and default
gradient component index is 0.

The graph is best visualized using a log-log scale.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6866 e5b18d87-469d-4833-9cc0-8cdfa06e9491