Add some error checks for files passed to convert4qmc.
Check that the files passed to '-gamessAscii' and '-ci' open without error
(that they exist and are readable).
Previously, if the file given to '-ci' did not exist, the code would enter
an infinite loop.

The reason for the hang is that the the 'lookFor' function in SimpleParser.h
only tests the 'eof' function, which will always be false for a non-existent
file. By contrast, the 'search' function tests the stream's bool overload,
and so will return on a bad file.  The problem could have been fixed here
(and lookFor still should be fixed), but it's good to have a more precise
error message (cannot open the file vs. cannot find a particular line in the
file.)


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@6962 e5b18d87-469d-4833-9cc0-8cdfa06e9491