Workaround for issue with Numpy 1.10 and greater.
The problem appears with a default install of Ubuntu 16.04.

Error message:
  File "/home/user/qmcpack/assembla/trunk/nexus/library/machines.py", line 1075, in submit_jobs
    if job.cores>self.cores and not nexus_core.generate_only:                
AttributeError: 'numpy.ndarray' object has no attribute 'cores'

I have not been able to create a simple reproducer that demonstrates the
problem.  It must be something with one the methods on the Job object 
(or its super classes) that causes numpy to convert that object to an array
of type bool.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@7072 e5b18d87-469d-4833-9cc0-8cdfa06e9491