Enable plane wave orbitals read from HDF files.
This should allow some cross checking of results from spline-based orbitals.
The plane wave version seems signicantly slower than splines.

Use 'pw', 'PW', or 'PWBasis' as the determinantset type.

One change from the existing plane wave code is the G vectors are not
re-truncated to a spherical shell.  The cutoff energy is not present in
the HDF file, which is necessary for this.

This code apparently uses a different (older?) method for creating single
particle orbitals. It does not work with the OrbitalImages estimator.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@7086 e5b18d87-469d-4833-9cc0-8cdfa06e9491