Add converter tests for #38. (The GAMESS CI converter failes for nact < 6)
Added two cases (the smallest, and one larger one), which should be enough to cover
the range of possibilities.

The command line needs the '-ci' option for this to work, so a feature was added
to read extra command line options from a 'cmd_args.txt' file (if present) in the
test case directory.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@7105 e5b18d87-469d-4833-9cc0-8cdfa06e9491