Fix flux estimator for complex wavefunctions.
Restore the test for solid LiH at an arbitrary k-point.  The reasoning in
revision 7113 is incorrect - need to use complex conjugate for one factor in
the integration by parts (or Green's identity), and then the exp(ikr) term
vanishes for the surface term.

With complex wavefunctions, the complex conjugate should be used for one
factor in the (grad psi)^2 term coming from the Green's identity.
However, regular complex multiplication should be used for the (grad psi)^2
term that comes from the converting the expression to use ln(psi) instead of psi.

Dot_CC and OTCDot_CC versions of the functions/templates have been added for
complex arguments.  These take the complex conjugate of the second argument
when computing the dot product.

Also fixed a bug in the D > 3 version of OTCDot.  Not sure any code actually
used this, though.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@7116 e5b18d87-469d-4833-9cc0-8cdfa06e9491