Compile formic libs under qmcpack.
Remove the post-build step and install step - these are not used by qmcpack.
And the post-build step is present, it fails because libformic_utils.a is not
where it expects it to be (qmcpack puts libraries in lib/ directory by default)

Fix MKL types to work (maybe an MKL version issue?)

Added hack to see if MKL is being used, don't want to work with non-MKL
blas/lapack configuration code right now.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/branches/new_optimizer@7182 e5b18d87-469d-4833-9cc0-8cdfa06e9491