Fix timer cache bug and unit tests.
Add the timer level to the unit tests.

As an optimization in previous check-in, the parent pointer is used as a
cache check - if the parent is the same as the top of the current timer stack,
the stack name does not need to be updated.
Need to invalidate this cache when the timer is popped off the stack.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/branches/timer@7339 e5b18d87-469d-4833-9cc0-8cdfa06e9491