Compile fixes for complex types.
Add multiplication operator for int and std::complex floating point types.
Add < operator to define an ordering for complex types.
Use specialization (and overloading) for the mynorm and myconj.
Add include of config.h in Utils.h so the correct ValueType is defined.

Note that complex wavefunctions still are not supported - it fails due to
assertions at runtime.


git-svn-id: https://subversion.assembla.com/svn/qmcdev/branches/afqmc@7370 e5b18d87-469d-4833-9cc0-8cdfa06e9491