Move non-git source version info to subdirectory
In the case where source is not a git repository, an empty repository info
file is created at configure time.  The git-rev.h file is placed in the src/
subdirectory.  If the code to create it is in the top-level CMakeLists.txt,
the src/ directory does not yet exist and it fails to create the file.
Move the git-ver.h file creation code to the src/CMakeLists.txt.