Soft link pseudopotential files in tests.
Some of the *.BFD.xml files in the test inputs are referenceed using a relative
symbolic link.  Using FILE(COPY ...) will keep the link, which then becomes a
dangling link pointing to nothing.  The previous method of
'cmake -E copy_directory' converted to link to an actual file copy.  The fix is
to treat the psuedopotential files the same as the *.h5 files, and the code
creates a symbolic link to the original file.

Also print a status message with the state of QMC_SYMLINK_TEST_FILES and a short
explanation of the consequences.