Limit the length of ParticleSet output.
As systems get larger, the particle position list in the output file gets
longer.  At some point, it really takes up too much space in a file that
should focus on human readability.
If needed, a record of the initial particle positions should be output some
other way - either a separate text file, an XML file, or an HDF file.