AFQMC - add sparse matrix multiply
Add routines for matrix-vector and matrix-matrix multiplication to remove the
hard dependence on MKL.   For performance on real calculations, a sparse matrix
library should still be used, but these routines are good enough for testing.

The gensparse.py script creates matrix multiplication test cases using numpy matrix
multiplication for comparison.  It's intended for use during development and
testing a new platform, but not necessarily for regular testing.