Add code coverage scripts for GCOV format
The code coverage model measures unit test coverage relative to a baseline
run (or runs).  There are three relevant runs:
  base - the baseline run (a typical run activating features to test against)
  unit - the unit test run
  diff - the results from the scripts comparing the coverage of unit relative to base

A test using a shortened diamondC_1x1x1_pp case is used for the base.
The added test has the label 'coverage'.  It should be possible to add more tests
with the 'coverage' label and automatically expand the baseline. (The compiler
coverage code automatically accumulates data from multiple runs)

Running with more than one MPI process will have race conditions when writing the coverage
data, and will likely corrupt it.  There might be ways to address this later.

Scripts in the tests/coverage directory:
 - read_gcov.py - read and write GCOV files
                - This can test that reading and writing is idempotent (produces the same file)
                  'python read_gcov.py --action test <gcov file>' will run that test.
 - demangle.py - call c++filt to demangle function names
 - compare_gcov.py - compare gcov files from two runs and produce the relative output
                   - It can also produce text output and summary of uncovered lines and
                      uncovered functions.  The output from this needs to be cleaned up more.
 - sample.cpp - a sample file for testing coverage scripts.  Directions in the code comments.

Still to come are scripts to integrate this with CTest.