Fix AFQMC unit test with Intel compiler.
The problem occurs on an empty 1x1 sparse matrix.

The problem seems to afflict the Intel compiler version 17.0.1 20161005 in release mode.
Any access to the A or M variables (to print them, for example) makes the problem disappear.
The workaround is to assign the array elements to temporary variables before testing them.