Initial tests for effective core potentials.
Move the Ylm.h file to Numerics and add some tests for it.
The default test checks a few values.  There is code that can test over
a wider range a values, but it is disabled by default.  This more extensive code
may be useful to run a new platform or under new conditions, but doesn't need to
be run as frequently as normal unit tests.  Eventually this kind of testing
should get moved to its own category.

The initial ECP tests use existing validation code for the spherical quadrature rules.
In order to make the function more testable, a return value is used to indicate
success or failure. The call to abort the program upon failure is moved higher
in the call chain.

Increase the tolerance of the check to allow the highest integration rule to pass.

The output to indicate the program in resize_warrays was removed - it doesn't
seem useful, and it clutters the output.

Add a few simple unit tests in Numerics for the vector sin/cos evaluation in e2iphi.h
and for the vector viewer.