Start moving optimizable determinant code across from Neuscamman group repo
On this branch, I will be moving across code that has already been
written on a separate and older repository, based on the original
SVN version of QMCPACK.

All of this code was written by Prof. Eric Neuscamman.

In this commit, the bulk of the wave function classes for the
optimizable determinants has been copied across, with some
initialization code moved across too. It compiles but doesn't
work yet, there is still more initialization code to be copied
across and corrected.