Update the code to use the putParametersInStandardForm function in place of resetParameters
putParametersInStandardForm is a wrapper function around
resetParameters, which was originally written with the optimizable
orbital code to be a more general function that allowed the orbital
rotations to be applied incrementally each update. It also includes
a bit of extra code to allow the updated orbital coefficients to
be printed out.

Since orbital rotations are now applied directly to the original
orbitals, instead of incrementally each update, it looks to me
like this may nto be necessary anymore, and we should maybe look
into moving back to the original function for better compatibility
with other bits of the code?