Copy across FDLR wave function class header file from SVN version of QMCPACK
This is code that was originally written in a separate version of
QMCPACK, which was based on the original SVN repo. I am now simply
moving it across to this newer github version of QMCPACK for all
future development and use.

This header contains the OrbitalBase derived class for treating
finite-difference linear response (FDLR) wave functions. This
wave function uses the linear response of a wave function with
orbital rotations to effectively allow a CIS wave function to
be treated also. The "derivative" of the linear response is
performed with a finite difference approximation, so that the
total wave function is simply a sum of two optimizable
determinants (i.e. with orbital rotations allowed).