Add insertFromSum and insertFromDiff members to the VariableSet struct
These functions are needed because the FDLR OrbitalBase object consists
of two TrialWavefunction obejcts. The first has parameters defined from
the sum of two sets of optimizable parameters, the second from the
difference of the optimizable parameter objects.

These functions allow these operations to be performed for the
optimizable parameter objects, in a way with a few safety checks
to make sure that equivalent parameters are being added.