Include contributions from the total wave function in the FDLR parameter derivatives
Before, these contributions were not being taken into account, i.e.
the quantities being calculated were the derivatives of H \psi/ \psi,
where \psi is *just* the FDLR wave function, not the entire wave
function, including the Jastrow factor and other potential factors.

This is now corected, which involves considering the gradient of
the whole wave function (P.G in the code), and needs the parameter
derivatives of the gradient of the log of the FDLR wave function,
which is now calculated. This is turn first requires the parameter
derivatives of the gradient of the log of the optimizable determinants,
hence the new member functions created in the last two commits.