Add unit tests for DMC walker propagation
Use some simplified wavefunctions for testing - constant orbital (gradient is zero)
and linear orbital (gradient is constant).

The methods in LinearOrbital contain abort statements so only the ones strictly
needed for the unit tests are added.  There is some implementation in 'evaluate',
but since it's not called by these tests, it may not be completely correct.

Move return statements in ConstantOrbital to the end of the function.