Coverage - mark lines not in unit tests correctly
For files that are completely untouched in the unit tests, but have some execution
in the the base tests, the behavior was to mark all the executed lines as uncovered.
This would incorrectly handle lines that are not executed (uncovered) in the base.
These lines should be marked as 'no-code' (not counted in coverage) in the output
directory.

This was most obvious in the case of formic/utils/lmyengine/engine_timing.cpp.
Even though the new optimizer is not called by the base test, there are some static
initializers that get called.   All the executable lines of the file were marked
as uncovered in the output, which was incorrect.  After the fix, only the lines
called by the static initializers are marked as uncovered.

With this fix the percentage coverage rises to 67% locally (probably will vary
some from that on the official test system)