Fix build if MPI is not found
Fixes issue #287 (Bug: Cmake configure does not complete without MPI due to test dependencies)

When MPI is not present, FindMPI sets the MPI_LIBRARY to the "not found" state,
causing a failure when using that library name.
The fix is to set MPI_LIBRARY to an empty string if MPI is not present.