Remove putParametersInStandardForm function from the code
This was previously added for LCOrbitalSetOpt, I believe to help allow
the orbital rotations to be applied incrementally on the previous
orbital rotation, rather than in total on the intial orbitals. It also
performed some printing of data. It was basically a wrapper function
around resetParameters.

The above idea is no longer performed or needed, so this commit
just reverts the code to using resetParameters always.

Note that there is no longer an option to print out orbitals,
as this would require an interface change to resetParameters. By
default it is turned off, since it is a lot of data, but the code
to turn it back on is still there.