Fix initialization of code so that it runs and gives the same results as before
The order in which things are initialized needed to be changed.
Previously, m_B was part of the LCOrbitalSetOptTrialFunc class.
LCOrbitalSetOpt had a pointer to this so could easily call it
when needed. Now everything from the TrialFunc class has been
moved to SlaterDetOpt, and it's not natural to call from the
SPOSetBase to the OrbitalBase object which holds it, so m_B
was moved to SPOSetBase object in the last commit. This commit
corrects initialization code to account for this change.