Change behavior for the DMC CPU time limit.
The time limit is set with the MaxCPUSecs parameter.

Previously, the time limit was measured from the beginning of the
DMC portion of the run, and continued until the set time runs out.
To effectively set this parameter, the user would need to know how much
time is spent prior to the DMC section, and the duration of each block,
in order to leave enough time for the simulation to shut down cleanly.

The new behavior measures the time elapsed from the beginning of the
run (captured during static initialization of the RunTimeManager object).
The time per DMC block is tracked and used to estimate if there is sufficient
time to run another block.

The time reserved to finish the next block is computed from the existing
average with some margin (to handle variations in block to block time) and
a fixed amount (to cover the time to end the simulation cleanly).