Add approximate testing for complex values.
Create a ComplexApprox class that acts similar to the Approx class in Catch.
It will compare the real and imaginary parts separately. It can also test just
the real part if compare_real_only is set.
The epsilon and scale values settings are not exposed, but it would be easy
to add them.
Also create a ValueApprox typedef for testing ValueType with the appropriate
real or complex setting.