Use the C MPI API in the formic library.
The C++ bindings to the MPI library were deprecated and are removed in MPI-3.
This change should enable QMCPACK to compile with newer versions of MPI
implementations.

Some issues:

There is no MPI_* type corresponding to MPI::Bool.  Hopefully translating to
MPI_INT should work okay.

MPI-3 added 'const' to a number of parameters in the C interface.
These 'const' qualifiers had already been present in the C++ interface.
When compiling with MPI-2.2 (or older), the formic wrappers use a
'const T*' to call a function with a non-const 'T *'.
Newer GCC versions (7.1) give an error about converting from 'const' to non-'const'.

The fix is to add const_cast's to remove the 'const' qualifier.