Add a basic H4 example with FRDL wave functions. No benchmarks yet, as optimization code is buggy.
I have not added benchmarks and the full test infrastructure,
as there appear to be errors in the optimization after the
changes to the SlaterDetOpt class. The FDLR commits have just
been cherry-picked from elsehwere onto here for the merge into
the develop branch on the main repo. This is just here as an
example to perhaps get review started for now.