Extract walker assignment code from walker swaps.
The walker swapping code first determines the new number of walkers each
processor should have, what exchanges are needed to get there, and then
communicates the walkers to be exchanged.
This change extracts the first two parts into a separate function, which
separates it from the MPI calls and makes it easier to test.

This commit also introduces a property-based test for this function, but
it is off by default.