Update in source comments.
Hit some issue on Mira when using >= 32768 ranks.

walker offsets (an array) of all the nodes are recorded in the config.h5 for restart.
It was written by every node to the file even though the file is open in collective IO mode.
It is not a problem when the node count was small (<=16384). With a large node count, hdf5 library crashed inside.
Changes to the offset IO:
write: collectively by every node, each contribute its own offset. The last node contributes 2 (for the end)
read: since the next run can be with a different node count, offset now is loaded by the master node and broadcast to all the nodes.
performance:
Both IO operations are measured faster.

Restart miniapp has passed all the checks with 65536 ranks.
restart tests also pass.