Update LogValue and PhaseValue in SlaterDet acceptMove
These weren't being updated for SlaterDet which was causing
problems for FDLR. TrialWaveFunction::acceptMove tries to
update LogValue, which it does incorrectly if SlaterDet doesn't
first update its LogValue in its call. This was causing problems
for FDLR since this updated value of LogValue is then used
when evalGrad is called. This presumably isn't the case for
other wave functions.