Hook up command line verbosity settings
Add the '-verbose' command line option, with values of 'low', 'high', or 'debug'.

Move the definition of the app_* streams from Configuration.h to OutputManager.h
(and change them to use the OutputManager so they respect the verbosity settings)

Set the default verbosity level to 'HIGH', to keep the output unchanged for now.