Switch timer summary output to use app_log
It used 'printf' previously and would not redirect timer output appropriately
with ensemble runs.
Given that the existing formatting works fine, and C++ stream formatting is
difficult to use, this check-in prints to a temporary buffer which is then
output to app_log.

Partially addresses #87.  The name of the info file still needs to be fixed.