Clean up random number and unit cell output
The random number info is not present in low-verbosity output.

The handling of wire/slab geometries resulted in double output of the unit cell info.
Now only do this printing for those boundary conditions.  Should also detect whether
the vacuum multiplier setting is different from 1.0, but that will require testing
to ensure it works.