Further cleanup of RNG and lattice output.
Update the RNG control output text.  Also print the offset in low verbosity mode.
The rest of the RNG seed generation is deterministic.
Remove output printing out seeds for individual processors and threads.

Skip printing the MPC density message if it's not being read.

Hard-code the lattice and volume units to 'bohr'.