Add unit tests for some factory classes
Testing for:
  HamiltonianPool
  HamiltonianFactory
  WaveFunctionFactory

The main purpose of this is setup of tests.  Actual testing of results is a little light.

The Libxml2Document forward declaration was moved out of the qmcplusplus namespace as the
Libxml2Document class is also outside that namespace.  If this not done, the use of
Libxml2Document in test_hamiltonian_pool will find the forward declaration, but not the
actual definition.