Add unit test for VMCSingleOMP
For ConstantOrbital.h - added one 'override', but then clang complained about all the
 other methods without it.  Using 'override' pointed out a virtual method that was no
 longer overriding anything - remove it.

Add diffuse.py for computing unit test values (the values were slightly different than
 the existing ones in test_vmc.cpp - update those to match.)
(This follows a general trend of putting the generic derivation/explanantion in
qmc_algorithms, and putting the particular python script for computing the unit test
values along side the unit tests.)

Starting point for tests for SimpleFixedNodeBranch, QMCUpdateBase, CloneManager.  The classes are
set up, but there are no tests there yet.

Added a constructor for initializing CloneManager to a number of threads specified by the test,
not just the number of OpenMP threads.