Add test for TraceManager.
It appears there exists test code to check compilation in TraceManager.cpp.
Move that to a unit test and fix it so it compiles and runs.

The operations in the first test may not be called in the correct order and may
not even be consistent.   They do run without crashing, though.