Fix names of scalar variable lists in tests
After fixing the QMC_RUN_AND_CHECK macro to ensure that the variable name
containing a list of scalar values actuall exists, these files showed up
as errors.  Fix them.

Note the fix for the AFQMC example only fixes the form of the call so the
right variable is passed in.  It still fails because 'Ebound' is not present
in the list of known scalars.