Change test comparison with reference value.
Use the combination of the error bar on the computed value and the reference
value for the comparison.  This change allows exact reference values (zero error),
like analytic results, or those coming from Gamess or Quantum Espresso.

For the case of the SHO VMC results, the exact value is known, and the local
energy is also a constant.   The averaging process introduces some numerical roundoff
error.  To account for this, increase the zero error reference by a little bit.

One issue is that a very noisy value can pass this test by having very large error bars.
The fix is to compare the variance in addition to the energy.

Add the variance to a couple of tests.