Add error bars to VMC SHO total energy
The errors are different with mixed vs. full precision.  The errors are numerical,
not statistical, but the checking script will still work.

Since the error check was added in the values, this commit reverts the changes in
check_scalars.py to add a small epsilon in the case the error is exactly zero.
The commits are 3af48cad1d91d6c431abc370e8199454d366ab20
and (partially) d0386ffd6eca4464ffbb7c0b97dd5739295a7187