Add unit test for GTO molecular orbitals.
Test case is a He atom with STO-3G basis.

The main body of the test is called twice - once to transform the GTO orbitals
to numerical orbitals (default QMCPACK behavior).  The other call evaluates the GTO form directly.
The XML is modified in memory to avoid keeping two slighly different versions of the input file.

The numerical orbital test fails in mixed precision.  The quintic splines lose too much precision
in the derivatives for small distances.  The test is excluded from mixed precision builds.
(Molecular calculations are usually done in full precision, so this is not too much of a restriction)