Extract 1D minimization routines from CuspCorr.cpp
There are two minimization loops in QMCWaveFunctions/MolecularOrbitals/CuspCorr.cpp.

This extracts the code so it can be tested separately and reused between the two loops.
The minimization occurs in two steps - first is to bracket the minimum, and the second is to perform
the actual minimization.

The two minimizations are slightly different in the bracketing step.  The second
loop (over a cutoff radius) includes some bounds (between 0 and some maximum value).  If the
max_bound value is set to a positive number, the code will check that the search does not go
outside the bounds.