Update gold files for converter tests
PR #1008 introduced the /application group to the HDF files.  This updates the gold files to match.

This problem was harder to find because h5diff prints no output when there are groups in one file
but not another.  The difference is visible with verbose (-v) output from h5diff.
Update the converter test script to output a message in this case.
(If this sort of failure comes up again, the script could be modified to always use -v with h5diff,
or rerun h5diff with -v if the comparison shows differences)