Start adding cusp correction pieces to SoA code.
Copy code for reading the cusp parameter files to the lcao/ directory.

Create routines for splitting out S-orbitals by modifying the MO coefficient matrix.
There are two of these routines.  One (splitEtaPhi) creates two wavefunctions - one with the
S-orbitals on a particular center, and the other wavefunction is everything else.  This is
used when generating the cusp correction.
(Routine was LCOrbitalSetWithCorrection::createLCOSets in AoS code.)

The other routine (removeSTypeOrbitals) removes S-orbitals on all centers and is the part
with no corrections.  This is used when evaluating the corrected wavefunction. The cusp
correction is evaluated separately and is added on.
(Routine was LCOrbitalSetWithCorrection::extractHighYLM in AoS code.)

In order to generalize this code between the two angular types (Cartesian and Spherical), some
methods have been added to find the S-type orbitals.  In particular, queryOrbitalForSType was
added to BasisSetBase.  Adding methods to base classes that are specific to a single
algorithm is less than ideal, but this seemed like the cleanest option.
Hopefully documenting which algorithm uses the method will help keep the confusion under control.

If a future algorithm should need to pick out specific orbitals, queryOrbitalsForSType could be
converted to a more general query (eg, accept a functor or something to describe the conditions)

The new files (CuspCorrection.h, CuspCorrection.cpp, and test_soa_cusp_corr.cpp)
have been run through clang-format.