Add radial evaluation for cusp correction
Update the python scripts to read the cusp correction parameters from the XML file
and generate the corrected radial values.

A similar unit test was written for the AoS version to ensure the python script is
correct, but it doesn't seem worth checking in.
Note the AoS version removes the angular normalization because the evaluation code puts it in.
The SoA version will not compute the angular part when evaluating the correction, so no need to remove it.