Connect cusp correction to wavefunction evaluation
Set up the correction in the builder.
Only write out the corrected orbital data files in debug mode.   These can be used to compare
against the AoS version.

This code does not correctly handle molecules with multiple atoms of the same type.
This will be fixed in a future commit.