Fix compile on BG/Q (Mira)
The required minimum cmake version also sets the policies according to that
cmake version.  The CMP0060 policy changed in cmake 3.3.

This policy change caused the libraries returned from finding the HDF library
to fail to link, because CMake previously altered the form on the link line,
and the new behavior does not.

The policy cannot be set in the toolchain file because of warnings related
to policy scope and CMP0011.