Remove Catch from NVCC-compiled code.
Reorganize the CUDA spline test to move the test case to a *.cpp file,
and keep only then minimal calls in the *.cu file.  This also means
performing all the Catch tests in the *.cpp file and not the *.cu file.

The Catch-specific REQUIRE call checking memory allocation was replaced
by an assert.  This should be good enough for now.  If more complete error
checking is desired, the return values from all the CUDA calls should be
checked as well.

This moves the Catch framework out of the *.cu file and avoids compilation
of it by nvcc.  Since Catch2 requires C++ 11 at minimum, this bypasses some
issues where the nvcc host compiler (eg, gcc 5) doesn't support C++ 11 by default.