Add unit test for Communicate class
Particularly the splitting constructor used for ensemble runs.
Test that it splits the processes as expected and creates the group leader communicator.
Purpose is to ensure this functionality doesn't regress when moving to the new MPI wrapper.

Add MPI initialization to the Catch MPI header.   The OHMMS::Controller initialization can
be removed from some of the unit tests that have it.

The test is designed to run under multiple MPI processes, but test system doesn't
handle this yet.  This test can be run manually to check that it works for different
numbers of MPI processes.